requirements.txt
Django>=4
django-extensions
djangorestframework
python-decouple==3.4
